Refactor: 배치작업 오류 처리 및 시간 당 조회수로 저장 로직으로 변경

+++ src/main/java/techit/gongsimchae/global/batch/views/ViewCountDto.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class ViewCountDto {
     private String subdivisionId;
-    private long viewCount;
+    private int viewCount;
 }
